Update README.md
At the two "Important" texts I was trying to create a nested list, but it is not displaying correctly when I preview the changes. Mike or Kevin, can you please fix this?
Also, for step 4 on Qubit 2.0, suggest the button for the app should be changed to "Celero PCR" rather than "NuGEN" so it is consistent across apps (and also allows for the addition of more NuGEN products in the future??).